main/requirements.txt
Flask>=2.0
python-dotenv>=1.0.0  # For loading environment variables (optional but recommended)
Flask-Session